pyproject.toml
[project]
name = "portfolio-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "agent-framework>=1.0.0b251120",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "yfinance>=1.2.0",
]
